Hiya UI cleanup
