+++ src/components/NavItems/tools/Loccus/components/HiyaHeader.jsx
@@ -1,0 +1,44 @@
+import React from "react";
+
+import Alert from "@mui/material/Alert";
+import Box from "@mui/material/Box";
+import Stack from "@mui/material/Stack";
+
+import { AudioFile } from "@mui/icons-material";
+
+import HeaderTool from "@Shared/HeaderTool/HeaderTool";
+import { i18nLoadNamespace } from "@Shared/Languages/i18nLoadNamespace";
+
+const HiyaHeader = () => {
+  const keyword = i18nLoadNamespace("components/NavItems/tools/Loccus");
+  const keywordAllTools = i18nLoadNamespace(
+    "components/NavItems/tools/Alltools",
+  );
+  const keywordWarning = i18nLoadNamespace("components/Shared/OnWarningInfo");
+
+  return (
+    <Box>
+      <HeaderTool
+        name={keywordAllTools("navbar_loccus")}
+        description={keywordAllTools("navbar_loccus_description")}
+        icon={
+          <AudioFile
+            style={{
+              fill: "var(--mui-palette-primary-main)",
+              height: "40px",
+              width: "auto",
+            }}
+          />
+        }
+      />
+      <Stack direction={"column"} spacing={2}>
+        <Alert severity="warning">
+          {keywordWarning("warning_beta_loccus")}
+        </Alert>
+        <Alert severity="info">{keyword("loccus_tip")}</Alert>
+      </Stack>
+    </Box>
+  );
+};
+
+export default HiyaHeader;
